requirements.txt
atproto
beautifulsoup4
feedparser
GitPython
filetype
Mastodon.py
pyyaml
tweepy
requests
twilio
